% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cardio.R
\name{model_cardio}
\alias{model_cardio}
\title{Model Cardio Function}
\usage{
model_cardio(age, sex, thalach)
}
\arguments{
\item{age}{edad del paciente}

\item{sex}{sexo del paciente ("Male","Female")}

\item{thalach}{presion sanguinea del paciente}
}
\value{

}
\description{
Model Cardio Function
}
\examples{
model_cardio(25,"Male",100)
}
